LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/font-relative-units-dynamic-expected.txt


FAIL em relative before change assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
FAIL rem relative before change assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
FAIL ex relative before change assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
FAIL ch relative before change assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
PASS em relative after change
PASS rem relative after change
PASS ex relative after change
PASS ch relative after change
